LangGraph_Course/first_examples/requirements.txt
langgraph
langchain_community
langchain
python-dotenv
ipykernel
deepinfra
langgraph-cli
langgraph-cli[inmem]
fastmcp
langchain-mcp-adapters